requirements.txt
apache-beam==2.28.0
avro-python3==1.9.2.1
beautifulsoup4==4.9.3
boto3==1.17.53
botocore==1.20.53
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
crcmod==1.7
Cython==0.29.23
dill==0.3.1.1
docopt==0.6.2
fastavro==1.4.0
future==0.18.2
google==3.0.0
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
hdfs==2.6.0
httplib2==0.17.4
idna==2.10
jmespath==0.10.0
mock==2.0.0
numpy==1.19.5
oauth2client==4.1.3
packaging==20.9
pbr==5.5.1
protobuf==3.15.8
psycopg2==2.8.6
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.3.7
six==1.15.0
soupsieve==2.2.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
